Client: make Clone create a new middleware stack
This allows an easier way of having Clients with a shared `HttpClient`,
but with potentially different middleware.

To clarify here: the middleware when cloned is still the same,
reference-based individual middleware, but the stack is new, so that
changes to the new client's stack are not reflected in any client it was
cloned from.

Closes https://github.com/http-rs/surf/issues/205